+++ src/inspect_ai/_eval/eval.py
@@ -4,9 +4,11 @@ import sys
 from pathlib import Path
 from typing import Any, Literal, cast
 
+from inspect_ai._eval.task.task import resolve_model_roles
 from inspect_ai._util.notgiven import NOT_GIVEN, NotGiven
 from inspect_ai.agent._agent import Agent, is_agent
 from inspect_ai.agent._as_solver import as_solver
+from inspect_ai.log._model import model_roles_config_to_model_roles
 
 if sys.version_info < (3, 11):
     from exceptiongroup import ExceptionGroup
@@ -70,6 +72,7 @@ def eval(
     model: str | Model | list[str] | list[Model] | None | NotGiven = NOT_GIVEN,
     model_base_url: str | None = None,
     model_args: dict[str, Any] | str = dict(),
+    model_roles: dict[str, str | Model] | None = None,
     task_args: dict[str, Any] | str = dict(),
     sandbox: SandboxEnvironmentType | None = None,
     sandbox_cleanup: bool | None = None,
@@ -116,6 +119,7 @@ def eval(
             with the model API.
         model_args: Model creation args
             (as a dictionary or as a path to a JSON or YAML config file)
+        model_roles: Named roles for use in `get_model()`.
         task_args: Task creation arguments
             (as a dictionary or as a path to a JSON or YAML config file)
         sandbox: Sandbox environment type
@@ -194,6 +198,7 @@ def eval(
                 model=model,
                 model_base_url=model_base_url,
                 model_args=model_args,
+                model_roles=model_roles,
                 task_args=task_args,
                 sandbox=sandbox,
                 sandbox_cleanup=sandbox_cleanup,
@@ -245,6 +250,7 @@ async def eval_async(
     model: str | Model | list[str] | list[Model] | None | NotGiven = NOT_GIVEN,
     model_base_url: str | None = None,
     model_args: dict[str, Any] | str = dict(),
+    model_roles: dict[str, str | Model] | None = None,
     task_args: dict[str, Any] | str = dict(),
     sandbox: SandboxEnvironmentType | None = None,
     sandbox_cleanup: bool | None = None,
@@ -286,7 +292,8 @@ async def eval_async(
             environment variable. Specify `None` to define no default model(s), which will
             leave model usage entirely up to tasks.
         model_base_url: Base URL for communicating with the model API.
-        model_args: Model creation args (as a dictionary or as a path to a JSON or YAML config file)
+        model_args: Model creation args (as a dictionary or as a path to a JSON or YAML config file
+        model_roles: Named roles for use in `get_model()`.
         task_args: Task creation arguments (as a dictionary or as a path to a JSON or YAML config file)
         sandbox: Sandbox environment type (or optionally a str or tuple with a shorthand spec)
         sandbox_cleanup: Cleanup sandbox environments after task completes (defaults to True)
@@ -368,7 +375,7 @@ async def eval_async(
 
         # resolve tasks
         resolved_tasks = eval_resolve_tasks(
-            tasks, task_args, model, GenerateConfig(**kwargs), sandbox
+            tasks, task_args, model, model_roles, GenerateConfig(**kwargs), sandbox
         )
 
         # warn and return empty string if we resolved no tasks
@@ -759,6 +766,9 @@ async def eval_retry_async(
             **eval_log.eval.model_args,
         )
 
+        # resolve model roles
+        model_roles = model_roles_config_to_model_roles(eval_log.eval.model_roles)
+
         # collect the rest of the params we need for the eval
         task_args = eval_log.eval.task_args
         tags = eval_log.eval.tags
@@ -815,9 +825,15 @@ async def eval_retry_async(
         log = (
             await eval_async(
                 tasks=PreviousTask(
-                    id=task_id, task=task, task_args=task_args, model=None, log=eval_log
+                    id=task_id,
+                    task=task,
+                    task_args=task_args,
+                    model=None,
+                    model_roles=None,
+                    log=eval_log,
                 ),
                 model=model,
+                model_roles=cast(dict[str, str | Model], model_roles),
                 task_args=task_args,
                 sandbox=eval_log.eval.sandbox,
                 sandbox_cleanup=sandbox_cleanup,
@@ -896,15 +912,19 @@ def eval_resolve_tasks(
     tasks: Tasks,
     task_args: dict[str, Any] | str,
     models: list[Model],
+    model_roles: dict[str, str | Model] | None,
     config: GenerateConfig,
     sandbox: SandboxEnvironmentType | None,
 ) -> list[ResolvedTask]:
+    resolved_model_roles = resolve_model_roles(model_roles)
     task_args = resolve_args(task_args)
     with task_display().suspend_task_app():
         resolved_tasks: list[ResolvedTask] = []
         for m in models:
             init_active_model(m, config)
-            resolved_tasks.extend(resolve_tasks(tasks, task_args, m, sandbox))
+            resolved_tasks.extend(
+                resolve_tasks(tasks, task_args, m, resolved_model_roles, sandbox)
+            )
         return resolved_tasks