+++ src/inspect_ai/log/_log.py
@@ -551,6 +551,22 @@ class EvalRevision(BaseModel):
     """Revision commit."""
 
 
+class EvalModelConfig(BaseModel):
+    """Model config."""
+
+    model: str
+    """Model name."""
+
+    config: GenerateConfig = Field(default_factory=GenerateConfig)
+    """Generate config"""
+
+    base_url: str | None = Field(default=None)
+    """Model base url."""
+
+    args: dict[str, Any] = Field(default_factory=dict)
+    """Model specific arguments."""
+
+
 class EvalSpec(BaseModel):
     """Eval target and configuration."""
 
@@ -607,6 +623,9 @@ class EvalSpec(BaseModel):
 
     model_args: dict[str, Any] = Field(default_factory=dict)
     """Model specific arguments."""
+
+    model_roles: dict[str, EvalModelConfig] | None = Field(default=None)
+    """Model roles."""
 
     config: EvalConfig
     """Configuration values for eval."""